修改字段数字类型

+++ manage/src/main/java/com/huanke/iot/manage/vo/response/device/customer/CustomerUserVo.java
@@ -28,6 +28,6 @@ public class CustomerUserVo {
     @Data
     public static class CustomerUserHourLiveCountVo {
         private String hour;
-        private BigDecimal userLiveCount;
+        private Integer userLiveCount;
     }
 }
